Module lta_models::utils::de

Deserialisation utils

Structs

WrapErr
Error for wrapped data

Traits

Sep
Separator trait

Functions

delimited
Errors

from_str
Generic implementation of FromStr

from_str_error_as_none
Errors

from_str_loc_to_loc
Errors

from_str_to_bool
Simple conversion of Y,Yes and N, No to boolean

from_str_to_coords
To be used when coordinates are space separated in a string and you would like to convert them to a Coordinates structure.

treat_error_as_none
Errors